refactor: Id generation reduced to single line and type of Id changed from Integer to Long

+++ src/main/java/com/codecool/budapestgo/dao/model/Stop.java
@@ -12,10 +12,8 @@ import lombok.*;
 @Entity
 public class Stop {
     @Id
-    @GeneratedValue(
-            strategy = GenerationType.SEQUENCE
-    )
-    private final Integer id;
+    @GeneratedValue(strategy = GenerationType.SEQUENCE)
+    private final Long id;
     @NonNull
     private String name;
     @Embedded
